Try resolving `size_t`.

+++ tests/fixtures/sizeof.rs
@@ -1,1 +1,1 @@
-pub const SIZEOF_INT: size_t = mem::size_of::<c_int>();
+pub const SIZEOF_INT: c_size_t = mem::size_of::<c_int>() as c_size_t;
